requirements.txt
appnope==0.1.4
asttokens==3.0.0
branca==0.8.2
certifi==2025.10.5
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
folium==0.20.0
fonttools==4.60.1
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.9.1
kafka-python==2.2.15
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
Pygments==2.19.2
pyparsing==3.2.5
pyspark==4.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
xyzservices==2025.4.0
